+++ angel_orm_generator/CHANGELOG.md
@@ -4,6 +4,7 @@
     * Stop assuming things have `id`, etc.
 * Resolve a bug where the `indexType` of `@Column` annotations. would not be found.
 * Add `cascade: true` to drops for hasOne/hasMany/ManyToMany migrations.
+* Support enum default values in migrations.
 
 # 2.0.0-dev.6
 * Fix bug where an extra field would be inserted into joins and botch the result.